Fully refactor web
